# docs/swagger.yaml
definitions:
...
  controller.ImportChannelFromOneAPIRequest:
    properties:
      dsn:
        type: string
    type: object
...
  middleware.APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
info:
  contact: {}
paths:
...
  /api/channels/import/oneapi:
    post:
      description: Imports channels from OneAPI
      parameters:
      - description: Import channel from OneAPI request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.ImportChannelFromOneAPIRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/middleware.APIResponse'
            - properties:
                data:
                  items: {}
                  type: array
              type: object
      security:
      - ApiKeyAuth: []
      summary: Import channel from OneAPI
      tags:
      - channels
...
swagger: "2.0"